requirements.txt
sqlalchemy>=2.0.0
boto3>=1.26.0
python-dotenv>=0.19.0
langchain>=0.0.350
langchain-openai>=0.0.5
fastapi>=0.95.0
requests>=2.31.0
crewai>=0.9.0
tiktoken<=0.5.2
boto3
python-dotenv
psycopg2-binary